Slight refactoring/rearrangement
