+++ shotgrid_leecher/repository/intermediate_hierarchy_repo.py
@@ -1,7 +1,7 @@
 from typing import List
 
 import shotgrid_leecher.utils.connectivity as conn
-from shotgrid_leecher.mapper import hierarchy_mapper
+from shotgrid_leecher.mapper import intermediate_mapper
 from shotgrid_leecher.record.enums import DbName
 from shotgrid_leecher.record.intermediate_structures import (
     IntermediateRow,
@@ -15,4 +15,4 @@ def fetch_by_project(
 ) -> List[IntermediateRow]:
     db = conn.get_db_client().get_database(DbName.INTERMEDIATE.value)
     result = db.get_collection(project_name).find({})
-    return [hierarchy_mapper.to_row(x, project_params) for x in result]
+    return [intermediate_mapper.to_row(x, project_params) for x in result]
