+++ tests/mapper/test_intermediate_mapper.py
@@ -3,7 +3,7 @@ import uuid
 from assertpy import assert_that
 from toolz import compose
 
-from shotgrid_leecher.mapper.hierarchy_mapper import to_shot
+from shotgrid_leecher.mapper.intermediate_mapper import to_shot
 from shotgrid_leecher.record.avalon_structures import AvalonProjectData
 from shotgrid_leecher.record.intermediate_structures import IntermediateShot
 from shotgrid_leecher.record.shotgrid_structures import (
